# perf: use `esbuild` to process `.mjs` files