--- a/e2e/path-mapping/__tests__/path-mapping.ts
+++ b/e2e/path-mapping/__tests__/path-mapping.ts
@@ -2,6 +2,7 @@ import { bar } from '@bar/bar-constant';
 
 test('path mapping should work without being affected by resolver', async () => {
   expect(bar).toBe(1);
+  // @ts-expect-error testing purpose
   await expect(import('@foo/foo-constant')).rejects.toMatchInlineSnapshot(
     `[Error: Cannot find module '@foo/foo-constant' from '__tests__/path-mapping.ts']`,
   );