--- a/e2e/transform-mjs/__tests__/transform-mjs.ts
+++ b/e2e/transform-mjs/__tests__/transform-mjs.ts
@@ -1,0 +1,6 @@
+// @ts-expect-error TypeScript < 4.5 doesn't support `.mjs` import
+import * as foo from '../foo.mjs';
+
+test('should transform mjs files', async () => {
+  expect(foo.pi).toBeDefined();
+});