--- a/e2e/transform-mjs/foo.mjs
+++ b/e2e/transform-mjs/foo.mjs
@@ -1,0 +1,3 @@
+const pi = 3.14
+
+export { pi }