--- a/e2e/transform-mjs/jest.config.js
+++ b/e2e/transform-mjs/jest.config.js
@@ -1,0 +1,12 @@
+/** @type {import('ts-jest/dist/types').ProjectConfigTsJest} */
+module.exports = {
+  displayName: 'transform-mjs',
+  globals: {
+    'ts-jest': {
+      tsconfig: '<rootDir>/../tsconfig.json',
+    },
+  },
+  transform: {
+    '^.+\\.(ts|js|mjs|html)$': '<rootDir>/../../build/index.js',
+  },
+};