--- a/package.json
+++ b/package.json
@@ -48,6 +48,7 @@
     "update-e2e": "node scripts/update-e2e-deps.js"
   },
   "dependencies": {
+    "esbuild": "0.13.12",
     "jest-environment-jsdom": "^27.0.0",
     "pretty-format": "^27.0.0",
     "ts-jest": "^27.0.0"