--- a/src/ng-jest-transformer.ts
+++ b/src/ng-jest-transformer.ts
@@ -2,6 +2,7 @@ import path from 'path';
 
 import type { TransformedSource } from '@jest/transform';
 import type { Config } from '@jest/types';
+import { transformSync } from 'esbuild';
 import { ConfigSet } from 'ts-jest/dist/config/config-set';
 import { TsJestTransformer } from 'ts-jest/dist/ts-jest-transformer';
 import type { ProjectConfigTsJest, TransformOptionsTsJest } from 'ts-jest/dist/types';
@@ -26,33 +27,19 @@ export class NgJestTransformer extends TsJestTransformer {
     const configSet = this._createConfigSet(transformOptions.config);
     /**
      * TypeScript < 4.5 doesn't support compiling `.mjs` file by default when running `tsc` which throws error
-     * ```
-     * File '/Users/ahn/ts-jest-only-example/foo.mjs' has an unsupported extension. The only supported extensions are
-     * '.ts', '.tsx', '.d.ts', '.js', '.jsx'.
-     * ```
-     * However, `transpileModule` API supports `.mjs` so we use it as a workaround.
      */
     if (path.extname(filePath) === '.mjs') {
-      const compilerOptions = configSet.parsedTsConfig.options;
-      const compilerModule = configSet.compilerModule;
-      const { outputText, sourceMapText, diagnostics } = compilerModule.transpileModule(fileContent, {
-        compilerOptions: {
-          ...compilerOptions,
-          module:
-            transformOptions.supportsStaticESM && configSet.useESM
-              ? compilerModule.ModuleKind.ES2020
-              : compilerModule.ModuleKind.CommonJS,
-        },
-        fileName: filePath,
-        reportDiagnostics: configSet.shouldReportDiagnostics(filePath),
+      const { target, sourceMap } = configSet.parsedTsConfig.options;
+      const { code, map } = transformSync(fileContent, {
+        loader: 'js',
+        format: 'cjs',
+        target: target === configSet.compilerModule.ScriptTarget.ES2015 ? 'es2015' : 'es2016',
+        sourcemap: sourceMap,
       });
-      if (diagnostics?.length) {
-        configSet.raiseDiagnostics(diagnostics, filePath);
-      }
 
       return {
-        code: outputText,
-        map: sourceMapText,
+        code,
+        map,
       };
     } else {
       return super.process(fileContent, filePath, transformOptions);