… src/__tests__/__snapshots__/ng-jest-transformer.spec.ts.snap
… yarn.lock