<!-- Thanks for submitting a pull request! Please provide enough information so that others can review your pull request. -->

## Summary

<!-- Explain the **motivation** for making this change. What existing problem does the pull request solve? -->
**With esbuild on Linux Node 12 no Jest cache**
```
transform /home/runner/work/jest-preset-angular/jest-preset-angular/examples/example-app-v13/node_modules/@angular/core/fesm2015/testing.mjs: 68.032ms
transform /home/runner/work/jest-preset-angular/jest-preset-angular/examples/example-app-v13/node_modules/@angular/core/fesm2015/core.mjs: 126.100ms
transform /home/runner/work/jest-preset-angular/jest-preset-angular/examples/example-app-v13/node_modules/@angular/compiler/fesm2015/compiler.mjs: 147.075ms
```

**With TypeScript transpileModule on Linux Node 12 no Jest cache**
```
transform /home/runner/work/jest-preset-angular/jest-preset-angular/examples/example-app-v13/node_modules/@angular/core/fesm2015/testing.mjs: 764.372ms
transform /home/runner/work/jest-preset-angular/jest-preset-angular/examples/example-app-v13/node_modules/@angular/core/fesm2015/core.mjs: 3948.006ms
transform /home/runner/work/jest-preset-angular/jest-preset-angular/examples/example-app-v13/node_modules/@angular/compiler/fesm2015/compiler.mjs: 4637.688ms
```

**With esbuild on Windows Node 12 no Jest cache**
```
transform D:\a\jest-preset-angular\jest-preset-angular\examples\example-app-v13\node_modules\@angular\core\fesm2015\testing.mjs: 105.483ms
transform D:\a\jest-preset-angular\jest-preset-angular\examples\example-app-v13\node_modules\@angular\core\fesm2015\core.mjs: 144.012ms
transform D:\a\jest-preset-angular\jest-preset-angular\examples\example-app-v13\node_modules\@angular\compiler\fesm2015\compiler.mjs: 176.672ms
```

**With TypeScript on Windows Node 12 no Jest cache**
```
transform D:\a\jest-preset-angular\jest-preset-angular\examples\example-app-v13\node_modules\@angular\core\fesm2015\testing.mjs: 1117.941ms
transform D:\a\jest-preset-angular\jest-preset-angular\examples\example-app-v13\node_modules\@angular\core\fesm2015\core.mjs: 5654.243ms
transform D:\a\jest-preset-angular\jest-preset-angular\examples\example-app-v13\node_modules\@angular\compiler\fesm2015\compiler.mjs: 6453.094ms
```

Closes #1141 

## Test plan

<!-- Demonstrate the code is solid. Example: The exact commands you ran and their output, screenshots / videos if the pull request changes UI. -->
Added unit test and e2e test

Green CI

## Does this PR introduce a breaking change?

- [ ] Yes
- [x] No

<!-- If this PR contains a breaking change, please describe the impact and migration plan -->


## Other information
**N.A.**